Adding different use cases
