+++ geapi/src/main/java/io/github/cdsap/geapi/ExperimentReport.kt
@@ -1,22 +1,17 @@
 package io.github.cdsap.geapi
 
-import io.github.cdsap.geapi.domain.GetMeasurements
-import io.github.cdsap.geapi.domain.impl.GetBuildScansWithQueryImpl
-import io.github.cdsap.geapi.domain.impl.GetMeasurementsImpl
-import io.github.cdsap.geapi.domain.impl.GetOutcomeReportImpl
-import io.github.cdsap.geapi.domain.impl.PrintExperimentResultsImpl
+import io.github.cdsap.geapi.domain.impl.*
 import io.github.cdsap.geapi.domain.model.Filter
 import io.github.cdsap.geapi.repository.impl.GradleRepositoryImpl
-import io.github.cdsap.geapi.view.OutcomeView
 
 class ExperimentReport(val filter: Filter, val repository: GradleRepositoryImpl) {
 
     suspend fun process() {
-        val getBuildScans = GetBuildScansWithQueryImpl(repository)
-        val getReport = PrintExperimentResultsImpl()
-        val measurements = GetMeasurementsImpl(repository)
-        val buildScansFiltered = getBuildScans.get(filter)
-        getReport.print(measurements.get(buildScansFiltered, filter))
+
+        val buildScansFiltered = GetBuildScansWithQueryImpl(repository).get(filter)
+        val buildsExperiment = FilterExperimentsImpl(repository).filter(buildScansFiltered, filter)
+
+        PrintExperimentResultsImpl().print(GetMeasurementsImpl(repository).get(buildsExperiment))
     }
 }
